a_Blog/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-ckeditor==6.5.1
django-debug-toolbar==3.7.0
django-js-asset==2.0.0
Pillow==9.2.0
python-dotenv==0.21.0
sqlparse==0.4.3
